Move Windows guest util functions to lib/windows
