+++ lib/windows/guest_tools.py
